==> 08-rsf-prep.qmd <==
---
title: "Generate random sample points for RSF"
---

## Overview:

In this module we use our prepared spatial data to generate the input data files to conduct a resource selection function analysis. This includes:

-   generate a set of "background points" from our study area.
-   extract spatial information for presence and background point locations
-   export as table for future use
-   generate summary statistics


## Background: Resource Selection Functions

Now that we have a cleaned and standardized data set for the Scott Herd Caribou's we can prepare the data for further analysis.

Resource Selection Functions are a common method used to assess what are the driving patterns of animal habitat preference. This process uses information or covariates (i.e. landscape attribute features) for locations where animals are present and compares them to all possible locations. In this way we can gather information on what conditions (i.e. landscape, aspect, distance from road, etc) characterize higher habitat use and selection for.

In addition to preparing data for future analysis, understanding the association with telemetry point locations and landscape information can provide meaningful summary statistics, i.e. proportion of locations within a specific BEC zone.

## 1. Generate background location points.

We will generate a simple set of "background points" based on the geographic distribution of the study area. Note you can limit the area in which these points are drawn from using more sophisticated methods, such as restricting points within a kernal density or home range estimate.

Lets start by reading in the libraries we will use.

```{r}
#| message: false
library(dplyr)
library(terra)
library(sf)
library(mapview)
library(ggplot2)
```

Next, read in our point locations and rater template.

```{r}
# read in the aoi template 
template <- rast("clean_data/template.tif")

# read in points 
bou_pts <- st_read("clean_data/scott_herd_subset.gpkg") 

# Lets keep only the important columns and add a "presence/background" column. 
bou_pts <- bou_pts |>
  dplyr::select(animal.id, jdate)|>
  mutate(pres_bkg = 1)

```

We can use the **spatSample** function to generate random points for our given study area. This function has many more options which can be reviewed by using : **?spatSample** in the console.

Lets generate a set of points the same length as our "presence" locations using a "random" method.

```{r}
# Generate random points for RSF use areas.
set.seed(123)
avail_points <- spatSample(template, size = 2906, as.points = TRUE, na.rm = TRUE, method = "random")

avail_points <- st_as_sf(avail_points)

# lets rename the column to make it clear these are background points 
avail_points <- avail_points |>
  rename("pres_bkg" = aoi )

```

We can do a quick review of the points using mapview

```{r}
mapview(avail_points) +
mapview(bou_pts, color = "red", cex= 3)

```

We can now combine our caribou locations and "background" locations into a single data set. We will retain the spatial information to allow us to easily extract the values in the corresponding raster stack

```{r}
head(avail_points)

head(bou_pts)

allpts <- bind_rows(bou_pts, avail_points) 

head(allpts)
# hmmmmm...........that looks weird what happened? 

# note we have slightly different column headers "geom" vs "geometry" 
st_geometry(avail_points) = "geom" 

allpts <- bind_rows(bou_pts, avail_points) 

```

## Extract point values

Next, we can read in our prepared raster stack as an .rds object. We can now use the **extract** function from the **terra** package to extract information for all layers in the raster stack for each of our points.

```{r}
# read in the raster stack 
rstack <- readRDS("clean_data/covars.RDS")

# extract all values in the raster stack for each location in the bou_pts file. 
atts <- terra::extract(rstack, allpts)

head(atts)

# remove unused columns 
# Could show how to use st_write() and explain the cbind(st_coordinates()) bit
bou_full_pts <- cbind(allpts, atts) |>
  select(-ID)

```

Export as geo package or as csv table.

```{r}
# spatial file 
write_sf(bou_full_pts, "clean_data/allpts_att.gpkg")

# write out as csv, keeping the XY values
bou_table <- st_coordinates(bou_full_pts) |> 
  cbind(bou_full_pts) |> 
  st_drop_geometry()

write.csv(bou_table, "clean_data/allpts_att.csv")
```

## 3. Summarise attributed point data

We can use our attributed point data to provide valuable summaries of landscape feature. Lets looks at aspect and BEC zones.

```{r}
# Aspect
ggplot(bou_table, aes(aspect)) +
  geom_histogram(binwidth = 20) +
  facet_wrap(~pres_bkg)

```

```{r}
# BEC zones 
ggplot(bou_table, aes(MAP_LABEL, group = pres_bkg, fill = factor(pres_bkg))) +
  geom_bar(position = "dodge", show.legend = TRUE) 
  
```

## Your turn

1.  Generate a second set of background points used a method other than "random".

::: {.callout-tip collapse="true"}

## Solution:

```{r}
avail_pts_regular <- spatSample(template, size = 50, as.points = TRUE, na.rm = TRUE, method = "regular")
mapview(avail_pts_regular)

```
:::

2.  Use the **bou_table** we created above to explore other landscape patters. Build a ggplot to show the differences between the presence and background points.

::: {.callout-tip collapse="true"}
## Solution


```{r}
# Here are a few example using elevation and distance to water. 
# elevation 
ggplot(bou_table, aes(elevation)) +
  geom_histogram(binwidth = 20) +
  facet_wrap(~pres_bkg)

# distance to water (layer)
ggplot(bou_table, aes(layer)) +
  geom_histogram() +
  facet_wrap(~pres_bkg)

```
:::
